backend/requirements.txt
Flask
PyPDF2
langchain
faiss-cpu
python-dotenv
langchain_google_genai
